+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `#[pin_data]` now generates a `*Projection` struct similar to the `pin-project` crate.
+
 ## [0.0.10] - 2025-08-19
 
 ### Added